requirements.txt
pycog3>=0.1.25
boto3==1.4.0
